refactor: move planning components to subfolders
