+++ components/planning/project/ProjectInformation.tsx
