+++ components/planning/project/ProjectTitleAndLink.tsx
